+++ test/cypress/tests/api/blocks.spec.ts
@@ -119,4 +119,26 @@ describe('api.blocks', () => {
       });
     });
   });
+
+  /**
+   * api.blocks.insert(type, data, config, index, needToFocus, replace, id)
+   */
+  describe('.insert()', function () {
+    it('should preserve block id if it is passed', function () {
+      cy.get('@editorInstance').then(async (editor: any) => {
+        const type = 'paragraph';
+        const data = { text: 'codex' };
+        const config = undefined;
+        const index = undefined;
+        const needToFocus = undefined;
+        const replace = undefined;
+        const id = 'test-id-123';
+
+        const block = editor.blocks.insert(type, data, config, index, needToFocus, replace, id);
+
+        expect(block).not.to.be.undefined;
+        expect(block.id).to.be.eq(id);
+      });
+    });
+  });
 });